+++ algorithms/Heap.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-void heapify(std::vector<double> &arr, int n, int i);
-void heapSort(std::vector<double> &arr);
+void heapify(std::vector<int> &arr, int n, int i);
+void heapSort(std::vector<int> &arr);
 
 #endif // HEAP_H
